+++ include/nfx/Graphics.h
@@ -13,6 +13,10 @@
 // Pipeline
 #include "nfx/graphics/gl/pipeline/RenderTarget.h"
 
+// Scene
+#include "nfx/graphics/gl/scene/cameras/Camera.h"
+#include "nfx/graphics/gl/scene/cameras/OrbitCamera.h"
+
 // Math
 #include "nfx/graphics/math/Mat4.h"
 #include "nfx/graphics/math/Quat.h"